Save changes to smart feeds
Previously `_feed_list_changed` wouldn't be set to true when the '(New
Headlines)' feed was added or when any smart feed was removed. This
lead to unintuitive behaviour where these changed wouldn't be saved if
they were the only changed made, but would be saved if other feeds were
changed as well. Thanks to @Uks2 for the fix.