Ignore atom <link> tags where rel != "alternate"
Previously, the atom parsing function would fill in an items link string
with the `href` attribute of the last `<link>` that it found. However,
because the spec (https://validator.w3.org/feed/docs/atom.html#link)
allows an item to have multiple `<link>` tags, with different meanings,
identified by the `rel` attribute, this can lead to the wrong url being
used.

The one we want, the link to the post content, should have the attribute
`rel="alternate"`, or, as this is the default, no `rel` attribute at
all. The atom parsing function now checks for this, and will not set an
items link string unless this is the case.

Thanks to @Uks2 for the fix.