Strip CDATA elements from content
Some incompetently written HTML descriptions contain CDATA tags,
sometimes in a horribly arranged soup that seems designed to break out
of a CDATA, which is what RSS wraps it with. The parser chokes on it, of
course, but once stripped, the item can be saved with only a few visual
artifacts of comments and tag pieces.