Encode feed title and url for urls.opml
It is not uncommon to have feeds with an ampersand in the URL or a
quotation mark in the title, which currently make snownews use a
malformed URL or not read the feed entry at all.

This probably also affects categories and filters, though because they
are user-defined, one can just not use special XML characters in them.